feat: moved user components to separate folder for organization
