+++ frontend/src/service/constant.ts
@@ -1,1 +1,1 @@
-export const API_URL = "atlas-uplb.vercel.app" as const;
+export const API_URL = "https://atlas-uplb.vercel.app" as const;
